#21 Using hooks to position first tab at top of t…
…he page. Not letting it overwrite other elements by specifying
z-index. Also renaming files.
